Добавил dev и production типы сборок. В dev отсылаются данные о крешах и аналитике, не выводятся логи. В production приложение подписывается ключом. Добавил два flavors: free и paid, в бесплатном из интерфейса убрал Избранное.
