+++ app/build.gradle
@@ -7,13 +7,21 @@ plugins {
 }
 
 android {
-    compileSdkVersion 30
+    signingConfigs {
+        KeyHW {
+            storeFile file('C:\\Users\\ivan\\.android\\keys\\ikorulev.jks')
+            storePassword 'Ngq7K93BmMuFL9ngpRsP'
+            keyPassword 'Ngq7K93BmMuFL9ngpRsP'
+            keyAlias 'KeyHW'
+        }
+    }
+    compileSdkVersion 31
     buildToolsVersion "30.0.3"
 
     defaultConfig {
         applicationId "ru.ikorulev.homework"
         minSdkVersion 21
-        targetSdkVersion 30
+        targetSdkVersion 31
         versionCode 1
         versionName "1.4"
 
@@ -21,50 +29,72 @@ android {
     }
 
     buildTypes {
+
         release {
+            signingConfig signingConfigs.KeyHW
+            resValue("bool", "FIREBASE_ANALYTICS_DEACTIVATED", "false")
+        }
+        debug {
+            resValue("bool", "FIREBASE_ANALYTICS_DEACTIVATED", "true")
+        }
+        dev {
             minifyEnabled false
             proguardFiles getDefaultProguardFile('proguard-android-optimize.txt'), 'proguard-rules.pro'
+            resValue("bool", "FIREBASE_ANALYTICS_DEACTIVATED", "true")
+        }
+        production {
+            minifyEnabled false
+            proguardFiles getDefaultProguardFile('proguard-android-optimize.txt'), 'proguard-rules.pro'
+            resValue("bool", "FIREBASE_ANALYTICS_DEACTIVATED", "false")
+            signingConfig signingConfigs.KeyHW
         }
     }
     compileOptions {
         sourceCompatibility JavaVersion.VERSION_1_8
         targetCompatibility JavaVersion.VERSION_1_8
     }
-    kotlinOptions {
-        jvmTarget = '1.8'
-    }
-    testOptions {
-        unitTests.returnDefaultValues = true
-        unitTests.includeAndroidResources = true
-
+    flavorDimensions 'app'
+    def MODE = "MODE"
+    productFlavors {
+        free {
+            dimension 'app'
+            buildConfigField "String", MODE , '\"free"'
+            applicationId "ru.ikorulev.homework.free"
+            signingConfig signingConfigs.KeyHW
+        }
+        paid {
+            dimension 'app'
+            buildConfigField "String", MODE , '\"paid"'
+            signingConfig signingConfigs.KeyHW
+        }
     }
 }
 
 dependencies {
 
     implementation "org.jetbrains.kotlin:kotlin-stdlib:$kotlin_version"
-    implementation 'androidx.core:core-ktx:1.3.2'
-    implementation 'androidx.appcompat:appcompat:1.2.0'
-    implementation 'com.google.android.material:material:1.3.0'
-    implementation 'androidx.constraintlayout:constraintlayout:2.0.4'
+    implementation 'androidx.core:core-ktx:1.6.0'
+    implementation 'androidx.appcompat:appcompat:1.3.1'
+    implementation 'com.google.android.material:material:1.4.0'
+    implementation 'androidx.constraintlayout:constraintlayout:2.1.0'
     implementation 'androidx.navigation:navigation-fragment-ktx:2.3.5'
     implementation 'androidx.navigation:navigation-ui-ktx:2.3.5'
     implementation "androidx.lifecycle:lifecycle-viewmodel-ktx:2.3.1"
-    implementation 'androidx.recyclerview:recyclerview:1.2.0'
+    implementation 'androidx.recyclerview:recyclerview:1.2.1'
     implementation 'androidx.lifecycle:lifecycle-extensions:2.2.0'
-    implementation "androidx.fragment:fragment-ktx:1.3.3"
+    implementation "androidx.fragment:fragment-ktx:1.3.6"
 
     implementation 'com.squareup.retrofit2:retrofit:2.9.0'
     implementation 'com.squareup.retrofit2:converter-gson:2.9.0'
     implementation 'com.squareup.retrofit2:adapter-rxjava2:2.9.0'
-    implementation "com.squareup.okhttp3:logging-interceptor:3.12.6"
-    implementation "com.squareup.okhttp3:okhttp-urlconnection:3.12.6"
-    implementation "com.github.bumptech.glide:glide:4.11.0"
+    implementation 'com.squareup.okhttp3:logging-interceptor:5.0.0-alpha.2'
+    implementation 'com.squareup.okhttp3:okhttp-urlconnection:5.0.0-alpha.2'
+    implementation 'com.github.bumptech.glide:glide:4.12.0'
 
-    implementation 'com.google.firebase:firebase-crashlytics:18.2.0'
-    implementation 'com.google.firebase:firebase-analytics:19.0.0'
-    implementation 'com.google.firebase:firebase-crashlytics-ktx:18.2.0'
-    implementation 'com.google.firebase:firebase-analytics-ktx:19.0.0'
+    implementation 'com.google.firebase:firebase-crashlytics:18.2.1'
+    implementation 'com.google.firebase:firebase-analytics:19.0.1'
+    implementation 'com.google.firebase:firebase-crashlytics-ktx:18.2.1'
+    implementation 'com.google.firebase:firebase-analytics-ktx:19.0.1'
     implementation 'com.google.firebase:firebase-messaging-ktx:22.0.0'
 
     testImplementation 'junit:junit:4.13.2'
@@ -74,7 +104,7 @@ dependencies {
     androidTestImplementation 'androidx.test.ext:junit:1.1.3'
     androidTestImplementation 'androidx.test.espresso:espresso-core:3.4.0'
     androidTestImplementation 'androidx.test.espresso:espresso-contrib:3.4.0'
-    androidTestImplementation 'org.mockito:mockito-android:2.28.2'
+    androidTestImplementation 'org.mockito:mockito-android:3.12.4'
 
     implementation "androidx.room:room-runtime:2.3.0"
     kapt "androidx.room:room-compiler:2.3.0"
@@ -86,8 +116,8 @@ dependencies {
     implementation 'io.reactivex.rxjava2:rxandroid:2.1.1'
     implementation "io.reactivex.rxjava2:rxkotlin:2.4.0"
 
-    implementation 'com.google.dagger:dagger:2.38'
-    kapt 'com.google.dagger:dagger-compiler:2.28.3'
-    kapt "com.google.dagger:dagger-android-processor:2.28.3"
+    implementation 'com.google.dagger:dagger:2.38.1'
+    kapt 'com.google.dagger:dagger-compiler:2.38.1'
+    kapt 'com.google.dagger:dagger-android-processor:2.38.1'
 
 }
