+++ app/free/production/output-metadata.json
@@ -1,0 +1,20 @@
+{
+  "version": 3,
+  "artifactType": {
+    "type": "APK",
+    "kind": "Directory"
+  },
+  "applicationId": "ru.ikorulev.homework.free",
+  "variantName": "freeProduction",
+  "elements": [
+    {
+      "type": "SINGLE",
+      "filters": [],
+      "attributes": [],
+      "versionCode": 1,
+      "versionName": "1.4",
+      "outputFile": "app-free-production.apk"
+    }
+  ],
+  "elementType": "File"
+}
