+++ app/src/free/res/menu/bottom_navigate.xml
@@ -1,0 +1,19 @@
+<?xml version="1.0" encoding="utf-8"?>
+<menu xmlns:android="http://schemas.android.com/apk/res/android">
+    <item
+        android:id="@+id/nav_list"
+        android:icon="@drawable/list"
+        android:title="@string/list" />
+    <item
+        android:id="@+id/nav_watch_later"
+        android:icon="@drawable/alarm"
+        android:title="@string/watch_later" />
+    <item
+        android:id="@+id/nav_download"
+        android:icon="@drawable/download"
+        android:title="@string/download" />
+    <item
+        android:id="@+id/nav_clear"
+        android:icon="@drawable/clear"
+        android:title="@string/clear" />
+</menu>
