--- app/src/main/res/layout/item_film.xml
@@ -1,73 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="wrap_content">
-
-    <ImageView
-        android:id="@+id/filmImage"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:layout_margin="16dp"
-        android:contentDescription="@string/filmImage"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent"
-        android:adjustViewBounds="true"
-        tools:src="@tools:sample/backgrounds/scenic"/>
-
-    <ImageView
-        android:id="@+id/favourites"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_marginStart="16dp"
-        android:src="@drawable/favourites_border"
-        android:contentDescription="@string/favourites"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintEnd_toStartOf="@id/watchLater"
-        app:layout_constraintHorizontal_chainStyle="spread_inside"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@id/filmImage" />
-
-    <ImageView
-        android:id="@+id/watchLater"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:contentDescription="@string/watch_later"
-        android:layout_marginStart="16dp"
-        android:src="@drawable/alarm"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintEnd_toStartOf="@id/filmTitle"
-        app:layout_constraintHorizontal_chainStyle="spread_inside"
-        app:layout_constraintStart_toEndOf="@id/favourites"
-        app:layout_constraintTop_toBottomOf="@id/filmImage" />
-
-    <TextView
-        android:id="@+id/filmTitle"
-        style="@android:style/TextAppearance.DeviceDefault.Medium"
-        android:layout_width="0dp"
-        android:layout_height="wrap_content"
-        android:gravity="start"
-        android:layout_margin="16dp"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintEnd_toStartOf="@id/filmButton"
-        app:layout_constraintStart_toEndOf="@id/watchLater"
-        app:layout_constraintTop_toBottomOf="@id/filmImage"
-        tools:text="@tools:sample/cities" />
-
-    <Button
-        android:id="@+id/filmButton"
-        style="@android:style/TextAppearance.DeviceDefault.Medium"
-        android:textColor="@color/white"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_marginEnd="16dp"
-        android:text="@string/detail"
-
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toEndOf="@id/filmTitle"
-        app:layout_constraintTop_toBottomOf="@id/filmImage" />
-
-
-</androidx.constraintlayout.widget.ConstraintLayout>
