+++ app/src/main/res/values/strings.xml
@@ -1,10 +1,4 @@
 <resources>
-    <string name="app_name">Finding movie</string>
-    <string name="hello">Hello</string>
-    <string name="title1">I Care a Lot</string>
-    <string name="title2">Daddy 2</string>
-    <string name="title3">Justice League</string>
-    <string name="title4">Shameless</string>
     <string name="detail">Details</string>
     <string name="inviteFriend">Invite a friend</string>
     <string name="messageText">I suggest we go to the movies</string>
@@ -15,11 +9,10 @@
     <string name="favourites">Favourites</string>
     <string name="list">List</string>
     <string name="download">Download</string>
-    <string name="Error">Loading error</string>
-    <string name="Retry">Retry the download?</string>
-    <string name="Clear">Clear tables</string>
+    <string name="error">Loading error</string>
+    <string name="retry">Retry the download?</string>
+    <string name="clear">Clear tables</string>
     <string name="watch_later">Watch later</string>
     <string name="select_date">Select the viewing date</string>
-    <string name="isFavorites" translatable="false">is favorites</string>
-    <string name="isNotFavorites" translatable="false">is not favorites</string>
+    <string name="hello" translatable="false">hello</string>
 </resources>
