+++ app/src/paid/res/values/strings.xml
@@ -1,0 +1,3 @@
+<resources>
+    <string name="app_name">Finding movies (paid)</string>
+</resources>
